Fix binary read version tag check
